+++ Woof/Woof/View/OwnerFlow/DetailedPetSitterView/DetailedPetSitterView.swift
@@ -8,15 +8,31 @@ struct DetailPetSitterView: View {
     var body: some View {
         VStack {
             AvatarView(url: viewModel.imageURL)
-            FiveStarRatingView(rating: viewModel.rating)
+            FiveStarRatingView(stars: viewModel.rating)
             TextWithIconLabelView(
                 iconName: .IconName.filledPerson,
                 text: viewModel.fullName
             )
+            .contextMenu {
+                Button {
+                    viewModel.copyToClipboardText(viewModel.fullName)
+                } label: {
+                    CopyToClipboardLabel()
+                }
+            }
+
             TextWithIconLabelView(
                 iconName: .IconName.phone,
                 text: viewModel.phoneNumber
             )
+            .contextMenu {
+                Button {
+                    viewModel.copyToClipboardText(viewModel.phoneNumber)
+                } label: {
+                    CopyToClipboardLabel()
+                }
+            }
+
             Text(viewModel.bio)
             Spacer()
         }