+++ Woof/Woof/ViewModel/DetailPetSitterViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import UIKit.UIPasteboard
 
 /// The view model for a detailed sitter card, responsible for preparing and providing data for the view.
 final class DetailSitterViewModel: ObservableObject {
@@ -8,8 +9,8 @@ final class DetailSitterViewModel: ObservableObject {
     /// The additional information about the sitter.
     @Published var bio: String
 
-    /// The rating of the sitter, on a scale from 0 to 5.
-    @Published var rating: Int
+    /// A 5-Star rating of the pet sitter.
+    @Published var rating: StarRating
 
     /// The phone number of the sitter.
     @Published var phoneNumber: String
@@ -23,10 +24,18 @@ final class DetailSitterViewModel: ObservableObject {
     init(sitter: Sitter) {
         self.sitter = sitter
         fullName = DataTransformer.fullName(name: sitter.name, surname: sitter.surname)
-        rating = DataTransformer.normalizeToZeroToFiveRange(sitter.rating)
+        rating = DataTransformer.fiveStarRating(for: sitter.rating)
         phoneNumber = sitter.phone
         imageURL = sitter.avatarUrl
         bio = sitter.bio
+    }
+
+    /// Copies the text to the clipboard. The copied text is available in the
+    /// systemwide general clipboard, which you use for general copy-paste operations.
+    ///
+    /// - Parameter text: The text to be copied to the clipboard.
+    func copyToClipboardText(_ text: String) {
+        UIPasteboard.general.string = text
     }
 
     private var sitter: Sitter